+++ CHANGELOG.md
@@ -15,9 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Documentation for Paper objects.
+- Path constructor typechecking input Vector array.
 
 ### Fixed
-
 
 ## [5.0.1]